Consolidate loggers per PR feedback
